Plug in Spreads interfaces as drop-in replacement for Ix ones
NB! Current Spreads contracts is broken, synchronous MoveNext will block if IAsyncEnumerable follows Spreads' behavior.
This is the last commit compatible with the original Ix. Going forward, conditional compilation expression will be used to check Spreads presence, and if not - define interfaces according to Spreads, but with covariance. Later we could use the project as a dependency for all other project if we need to keep covariance or need some C# even for collections (e.g. Hopac.Core uses C# because F# is sometimes slow)